+++ apps/backend/src/modules/user-routes/libs/types/type.ts
@@ -1,5 +1,6 @@
 export { type UserRouteFilter } from "./user-route.filter.type.js";
 export {
+	type UserRouteDeleteParameters,
 	type UserRouteFinishRequestDto,
 	type UserRouteQueryRequestDto,
 	type UserRouteResponseDto,