+++ packages/shared/src/modules/routes/libs/types/route-find-all-options.type.ts
@@ -1,4 +1,6 @@
-type RouteFindAllOptions = {
+import { type PaginationQuery } from "../../../../libs/types/types.js";
+
+type RouteFindAllOptions = PaginationQuery & {
 	categories?: string | string[];
 	latitude?: number;
 	longitude?: number;