+++ packages/shared/src/modules/routes/libs/types/route-get-by-id-response-dto.type.ts
@@ -1,9 +1,11 @@
 import { type LineStringGeometry } from "../../../../libs/types/types.js";
 import { type FileUploadResponseDto } from "../../../files/libs/types/types.js";
+import { type UserRouteResponseDto } from "../../../user-routes/user-routes.js";
 
 type RouteGetByIdResponseDto = {
+	createdAt: string;
 	createdByUserId: number;
-	description: string;
+	description: null | string;
 	distance: number;
 	duration: number;
 	geometry: LineStringGeometry;
@@ -15,6 +17,7 @@ type RouteGetByIdResponseDto = {
 		name: string;
 		visitOrder: number;
 	}[];
+	savedUserRoute: null | Pick<UserRouteResponseDto, "id" | "status">;
 };
 
 export { type RouteGetByIdResponseDto };