+++ packages/shared/src/modules/user-routes/libs/types/user-route-delete-parameters-request.type.ts
@@ -1,0 +1,5 @@
+type UserRouteDeleteParameters = {
+	id: number;
+};
+
+export { type UserRouteDeleteParameters };